# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Data Analyst Agent API
  version: 1.0.0
  description: |
    AI Data Analyst Agent 前后端共享契约。
    具体通用规则见 API_INTEGRATION_STANDARD.md。
  contact:
    name: AI Data Analyst Agent 项目组
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发
  - url: /api/v1
    description: 同源部署
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 使用部署账号创建登录会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Error"
  /auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      summary: 获取当前登录会话
      responses:
        "200":
          description: 当前会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Error"
...
  /projects/{project_id}/analysis-specs:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Analysis]
      operationId: listAnalysisSpecs
      summary: 查询分析定义
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: dataset_version_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 分析定义列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisSpecPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Analysis]
      operationId: createAnalysisSpec
      summary: 创建分析定义草稿
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisSpecCreate"
      responses:
        "201":
          description: 分析定义
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisSpec"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/analysis-specs/{spec_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/SpecId"
    get:
      tags: [Analysis]
      operationId: getAnalysisSpec
      summary: 获取分析定义
      responses:
        "200":
          description: 分析定义
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisSpec"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Analysis]
      operationId: updateAnalysisSpec
      summary: 编辑分析定义草稿并创建新修订
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisSpecUpdate"
      responses:
        "200":
          description: 新修订
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisSpec"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/analysis-specs/{spec_id}/confirm:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/SpecId"
    post:
      tags: [Analysis]
      operationId: confirmAnalysisSpec
      summary: 确认分析定义
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 已确认的分析定义
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisSpec"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: datatrace_session
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^prj_[A-Za-z0-9]+$"
...
    SpecId:
      name: spec_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^spec_[A-Za-z0-9]+$"
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 资源 revision，格式为双引号包围的整数
      schema:
        type: string
        example: '"3"'
  responses:
    Unauthorized:
      description: 登录凭证缺失、无效或已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error:
      description: 错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在或无权访问
              request_id: req_01JABC
              retryable: false
              details: {}
    Conflict:
      description: 当前状态或重复请求冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PreconditionFailed:
      description: 客户端资源版本已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VERSION_CONFLICT
              message: 资源已被修改，请刷新后重试
              request_id: req_01JABC
              retryable: false
              details:
                current_revision: 4
    ValidationError:
      description: 业务校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 120
        password:
          type: string
          minLength: 1
          maxLength: 500
          format: password
    SessionInfo:
      type: object
      required: [subject_id]
      properties:
        subject_id:
          type: string
        expires_in_seconds:
          type: [integer, "null"]
          minimum: 1
...
    AnalysisSpecBase:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        dataset_version_id:
          type: string
        task:
          type: string
          enum: [descriptive, comparison, statistical_test, binary_classification, multiclass_classification, regression]
        target:
          type: [string, "null"]
        entity_key:
          type: [string, "null"]
        time_column:
          type: [string, "null"]
        prediction_time_description:
          type: [string, "null"]
          maxLength: 1000
        split_strategy:
          type: string
          enum: [none, random, stratified, temporal, group]
        group_column:
          type: [string, "null"]
        metrics:
          type: array
          items:
            type: string
        included_columns:
          type: array
          items:
            type: string
        excluded_columns:
          type: array
          items:
            type: string
        random_seed:
          type: integer
          default: 42
        causal_interpretation_allowed:
          type: boolean
          const: false
    AnalysisSpecCreate:
      allOf:
        - $ref: "#/components/schemas/AnalysisSpecBase"
        - type: object
          required: [name, dataset_version_id, task, split_strategy, metrics]
    AnalysisSpecUpdate:
      allOf:
        - $ref: "#/components/schemas/AnalysisSpecBase"
    AnalysisSpec:
      allOf:
        - $ref: "#/components/schemas/AnalysisSpecBase"
        - type: object
          required:
            - spec_id
            - project_id
            - revision
            - status
            - name
            - dataset_version_id
            - task
            - split_strategy
            - metrics
            - created_at
            - updated_at
          properties:
            spec_id:
              type: string
              example: spec_01JABC
            project_id:
              type: string
            revision:
              type: integer
              minimum: 1
            status:
              type: string
              enum: [draft, confirmed, superseded]
            validation_warnings:
              type: array
              items:
                type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    AnalysisSpecPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AnalysisSpec"
...
    PageMeta:
      type: object
      required: [page, page_size, total, has_more]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message, request_id, retryable, details]
      properties:
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true